requirements.txt
allure-pytest==2.9.45
pytest-xdist==2.5.0
pytest==7.1.1
requests==2.27.1
selenium==4.1.3
webdriver_manager==3.5.4
Faker==13.12.0